pyproject.toml
[tool.poetry]
name = "async-sqlite-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

aioquic = "^1.1.0"
deprecation = "^2.1.0"
pyyaml = "^6.0.1"
fastapi = "^0.111.1"
hypercorn = "^0.17.3"
aiosqlite = "^0.20.0"
aiofiles = "^24.1.0"
nuitka = "^2.3.11"
pynng = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
